requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
discord.py==2.4.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pillow==10.4.0
praw==7.7.1
prawcore==2.4.0
python-dotenv==1.0.1
requests==2.32.3
update-checker==0.18.0
urllib3==2.2.2
websocket-client==1.8.0
yarl==1.9.4
